requirements.txt
appnope==0.1.0
attrs==18.2.0
backcall==0.1.0
bleach==3.0.2
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
GPy==1.9.5
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.3
pandas==0.23.4
pandocfilters==1.4.2
paramz==0.9.4
parso==0.3.1
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.4.2
prompt-toolkit==2.0.6
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.5
pytz==2018.7
pyzmq==17.1.2
qtconsole==4.4.2
scipy==1.1.0
Send2Trash==1.5.0
six==1.11.0
statsmodels==0.9.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
